♻️ Refactor: 베럴 파일에 컴포넌트 추가 및 적용 (#25)
