+++ src/main.tsx
@@ -4,7 +4,7 @@ import { createRoot } from 'react-dom/client'
 
 import '@/index.css'
 import App from '@/App.tsx'
-import ToastContainer from '@/components/common/Toast/ToastContainer'
+import { ToastContainer } from '@/components'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
