+++ src/store/index.ts
@@ -1,1 +1,4 @@
 // store
+import { useToastStore } from '@/store/useToastStore'
+
+export { useToastStore }
