+++ src/types/index.ts
@@ -1,1 +1,4 @@
 // types
+import type { Toast } from '@/types/Toast'
+
+export type { Toast }
